move LD client init options+user obj to constants; move "ready" handler to handlers.ts
